media-source/SourceBuffer-appendWindowEnd_removed.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Setting SourceBuffer#appendWindowEnd after the sourceBuffer object is removed</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var video = document.querySelector('video');
var mimes = ['video/webm; codecs="vorbis,vp8"', 'video/mp4'];

//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
async_test(function(t) {
    function test(mime, done) {
        if(!canPlayType.bind(t)(mime)) {
            return;
        }
        var mediaSource = new MediaSource();
        mediaSource.addEventListener('sourceopen', function(e) {
            var sourceBuffer = this.addSourceBuffer(mime);
            this.removeSourceBuffer(sourceBuffer);
            t.step(function() {
                assert_throws({name: 'InvalidStateError'},
                              function() {
                                  sourceBuffer.appendWindowEnd = 1;
                              });
            });
            if(done) {
                t.done();
            }
        });
        video.src = window.URL.createObjectURL(mediaSource);
    }
    //execute this test for videos and corresponding MIME types
    mimes.forEach(function(mime, i) {
        test(mimes[i], i == mimes.length - 1);
    });
},
'SourceBuffer#appendWindowEnd : After this object is removed from the' +
'sourceBuffers attribute of the parent media source, INVALID_STATE_ERR exception should be thrown.');
</script>
</body>
</html>
